moved test/TTFontTest to utils::TTFont. updated SpriteText. still fixing EasingEquations.
git-svn-id: https://shooting-cubes.googlecode.com/svn@107 93678846-683e-0410-93c7-9b11d4a5ef47

+++ trunk/include/TextureAnimator.hpp
@@ -36,7 +36,7 @@ public:
 
     virtual void animateNode(ISceneNode* node, u32 timeMs) {
         if ( !node ) return;
-        if ( timeMs < this->startTime_ ) return;
+        if ( static_cast<s32>(timeMs) < this->startTime_ ) return;
         if ( animation_.size() < 1 ) return;
 
         u32 t = ( timeMs - this->startTime_ );
